chore: Reestructure the code: Separates the use-style hook into a diferent file, and improve the legibility of the onChange method
